requirements.txt
python>=3.7.4
numpy>=1.17.2
paddlepaddle-gpu>=1.7.0
visualdl==1.3.0